move latin hypercube generator back
