+++ xopt/generators/__init__.py
@@ -59,7 +59,7 @@ def get_generator_dynamic(name: str) -> type[Generator]:
             )
     elif name in all_generator_names["scipy"]:
         try:
-            from xopt.generators.latin_hypercube import LatinHypercubeGenerator
+            from xopt.generators.scipy.latin_hypercube import LatinHypercubeGenerator
             from xopt.generators.sequential.neldermead import NelderMeadGenerator
             from xopt.generators.sequential.scipy import ScipyGenerator
