+++ xopt/generators/scipy/latin_hypercube.py
